now iterates shredding each successive page
previously it used to shred just a static page, it is now capable for
_recursively_ requesting a successive page until it can no longer find
classifieds.

note: this was suprisingly harder to do than expected due to the
non-blocking async nature of node.js.  decided to use a recursive
approach due to the underdeterministic # of requests that is expected
to take place (http://stackoverflow.com/a/6048595)